renaming files and added plots
