+++ selection/sampling/randomized/tests/test_lasso_randomX_langevin.py
@@ -3,9 +3,10 @@ from scipy.stats import laplace, probplot, uniform
 
 from selection.algorithms.lasso import instance
 import selection.sampling.randomized.api as randomized
-from pvalues1 import pval
+from pvalues_randomX import pval
 from matplotlib import pyplot as plt
 import regreg.api as rr
+import selection.sampling.randomized.losses.lasso_randomX as lasso_randomX
 
 
 def test_lasso(s=5, n=100, p=20):
@@ -15,7 +16,7 @@ def test_lasso(s=5, n=100, p=20):
     lam_frac = 1.
 
     randomization = laplace(loc=0, scale=1.)
-    loss = randomized.gaussian_Xfixed(X, y)
+    loss = lasso_randomX.lasso_randomX(X, y)
 
     random_Z = randomization.rvs(p)
     epsilon = 1.
@@ -39,9 +40,9 @@ def test_lasso(s=5, n=100, p=20):
     # data, opt_vars = sampler1.state
 
     # initial solution
+
     problem = rr.simple_problem(loss, penalty)
-    random_term = rr.identity_quadratic(epsilon, 0,
-                                        random_Z, 0)
+    random_term = rr.identity_quadratic(epsilon, 0, random_Z, 0)
     solve_args = {'tol': 1.e-10, 'min_its': 100, 'max_its': 500}
     initial_soln = problem.solve(random_term, **solve_args)
     initial_grad = loss.smooth_objective(initial_soln,  mode='grad')
@@ -50,13 +51,32 @@ def test_lasso(s=5, n=100, p=20):
                                          random_Z,
                                          epsilon)
 
-    data = y.copy()
     active = penalty.active_set
     inactive = ~active
-
-    #betaE, cube = opt_vars
+    loss.fit_E(active)
+    beta_unpenalized = loss._beta_unpenalized
+    residual = y - np.dot(X[:, active], beta_unpenalized)  # y-X_E\bar{\beta}^E
+    N = np.dot(X[:, inactive].T, residual)  # X_{-E}^T(y-X_E\bar{\beta}_E), null statistic
+    data = np.concatenate((beta_unpenalized, N), axis=0)
     ndata = data.shape[0];  nactive = betaE.shape[0];  ninactive = cube.shape[0]
-    init_vec_state = np.zeros(n+nactive+ninactive)
+
+
+    # parametric coveriance estimate
+    XE_pinv = np.linalg.pinv(X[:, active])
+    mat = np.zeros((nactive+ninactive, n))
+    mat[:nactive,:] = XE_pinv
+    mat[nactive:,:] = X[:, inactive].T.dot(np.identity(n)-X[:, active].dot(XE_pinv))
+
+    Sigma_full = mat.dot(mat.T)
+    Sigma_full_inv= np.linalg.inv(Sigma_full)
+
+
+    # non-parametric covariance estimate
+    #Sigma_full = loss._Sigma_full
+    #Sigma_full_inv = np.linalg.inv(Sigma_full)
+
+
+    init_vec_state = np.zeros(ndata+nactive+ninactive)
     init_vec_state[:ndata] = data
     init_vec_state[ndata:(ndata+nactive)] = betaE
     init_vec_state[(ndata+nactive):] = cube
@@ -69,6 +89,7 @@ def test_lasso(s=5, n=100, p=20):
         cube = vec_state[(ndata+nactive):]
 
         signs = penalty.signs
+
         projected_betaE = betaE.copy()
         projected_cube = np.zeros_like(cube)
 
@@ -81,7 +102,7 @@ def test_lasso(s=5, n=100, p=20):
         return np.concatenate((data, projected_betaE, projected_cube), 0)
 
 
-    def full_gradient(vec_state, loss=loss, penalty =penalty, X=X,
+    def full_gradient(vec_state, loss=loss, penalty = penalty, Sigma_full_inv=Sigma_full_inv,
                       lam=lam, epsilon=epsilon, ndata=ndata, active=active, inactive=inactive):
         nactive = np.sum(active); ninactive=np.sum(inactive)
 
@@ -93,7 +114,7 @@ def test_lasso(s=5, n=100, p=20):
         params , _ , opt_vec = penalty.form_optimization_vector(opt_vars) # opt_vec=\epsilon(\beta 0)+u, u=\grad P(\beta), P penalty
 
         gradient = loss.gradient(data, params)
-        hessian = loss.hessian()
+        hessian = loss.hessian
 
         ndata = data.shape[0]
         nactive = betaE.shape[0]
@@ -101,33 +122,41 @@ def test_lasso(s=5, n=100, p=20):
 
         sign_vec = - np.sign(gradient + opt_vec)  # sign(w), w=grad+\epsilon*beta+lambda*u
 
-        B = hessian + epsilon * np.identity(nactive + ninactive)
-        A = B[:, active]
+        A = hessian + epsilon * np.identity(nactive + ninactive)
+        A_restricted = A[:, active]
 
+        T = data[:nactive]
         _gradient = np.zeros(ndata + nactive + ninactive)
-        _gradient[:ndata] = - (data + np.dot(X, sign_vec))
-        _gradient[ndata:(ndata + nactive)] = np.dot(A.T, sign_vec)
-        _gradient[(ndata + nactive):] = lam * sign_vec[inactive]
 
-        # data_proposal = np.dot(P, data) + np.dot(R, data_proposal)
+        # saturated model
+        _gradient[:ndata] = - np.dot(Sigma_full_inv, data)
+        _gradient[:nactive] -= hessian[:,active].T.dot(sign_vec)
+        _gradient[nactive:(ndata)] -= sign_vec[inactive]
+
+        # selected model
+        #_gradient[:nactive] = - (np.dot(Sigma_T_inv, data[:nactive]) + np.dot(hessian[:, active].T, sign_vec))
+        _gradient[ndata:(ndata + nactive)] = np.dot(A_restricted.T, sign_vec)
+        _gradient[(ndata + nactive):] = lam * sign_vec[inactive]
 
         return _gradient
 
 
     null, alt = pval(init_vec_state, full_gradient, full_projection,
-                      X, y, nonzero, active)
+                      Sigma_full[:nactive, :nactive], data, nonzero, active)
 
     return null, alt
 
 if __name__ == "__main__":
 
     P0, PA = [], []
-    for i in range(50):
+    for i in range(20):
         print "iteration", i
         p0, pA = test_lasso()
         P0.extend(p0); PA.extend(pA)
 
+    print "done! mean: ", np.mean(P0), "std: ", np.std(P0)
     plt.figure()
     probplot(P0, dist=uniform, sparams=(0,1), plot=plt, fit=True)
-    plt.plot([0,1], color='k', linestyle='-', linewidth=2)
+    plt.plot([0, 1], color='k', linestyle='-', linewidth=2)
+    plt.suptitle("LASSO with random X")
     plt.show()
